hosted/cmsis_dap: Fixed a couple of clang-tidy lints in dbg_dap_cmd_bulk()

+++ src/platforms/hosted/cmsis_dap.c
@@ -333,7 +333,7 @@ ssize_t dbg_dap_cmd_bulk(const uint8_t *const request_data, const size_t request
 {
 	int transferred = 0;
 	const int result = libusb_bulk_transfer(
-		usb_handle, out_ep, (uint8_t *)request_data, request_length, &transferred, TRANSFER_TIMEOUT_MS);
+		usb_handle, out_ep, (uint8_t *)request_data, (int)request_length, &transferred, TRANSFER_TIMEOUT_MS);
 	if (result < 0) {
 		DEBUG_WARN("CMSIS-DAP write error: %s (%d)\n", libusb_strerror(result), result);
 		return result;
@@ -341,8 +341,8 @@ ssize_t dbg_dap_cmd_bulk(const uint8_t *const request_data, const size_t request
 
 	/* We repeat the read in case we're out of step with the transmitter */
 	do {
-		const int result =
-			libusb_bulk_transfer(usb_handle, in_ep, response_data, response_length, &transferred, TRANSFER_TIMEOUT_MS);
+		const int result = libusb_bulk_transfer(
+			usb_handle, in_ep, response_data, (int)response_length, &transferred, TRANSFER_TIMEOUT_MS);
 		if (result < 0) {
 			DEBUG_WARN("CMSIS-DAP read error: %s (%d)\n", libusb_strerror(result), result);
 			return result;
